refactor(core): move standalone canary analysis domain models back into it's module (#742)
