+++ kayenta-standalone-canary-analysis/src/main/java/com/netflix/kayenta/standalonecanaryanalysis/CanaryAnalysisConfig.java
@@ -17,7 +17,7 @@
 package com.netflix.kayenta.standalonecanaryanalysis;
 
 import com.netflix.kayenta.canary.CanaryConfig;
-import com.netflix.kayenta.domain.standalonecanaryanalysis.CanaryAnalysisExecutionRequest;
+import com.netflix.kayenta.standalonecanaryanalysis.domain.CanaryAnalysisExecutionRequest;
 import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
