+++ kayenta-standalone-canary-analysis/src/main/java/com/netflix/kayenta/standalonecanaryanalysis/storage/StandaloneCanaryAnalysisObjectType.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2020 Playtika.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.netflix.kayenta.standalonecanaryanalysis.storage;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.netflix.kayenta.standalonecanaryanalysis.domain.CanaryAnalysisExecutionStatusResponse;
+import com.netflix.kayenta.storage.ObjectType;
+import com.netflix.kayenta.storage.StandardObjectType;
+import lombok.experimental.UtilityClass;
+
+@UtilityClass
+public class StandaloneCanaryAnalysisObjectType {
+
+  public static final ObjectType STANDALONE_CANARY_RESULT_ARCHIVE =
+      new StandardObjectType(
+          new TypeReference<CanaryAnalysisExecutionStatusResponse>() {},
+          "standalone_canary_archive",
+          "standalone_canary_archive.json");
+}
